+++ src/main/java/com/example/app_fast_food/product/dto/ProductResponseDTO.java
@@ -52,5 +52,5 @@ public class ProductResponseDTO {
     private List<DiscountResponseDTO> discounts;
     private List<BonusResponseDTO> bonuses;
 
-    private Double finalPrice;
+    private Long finalPrice;
 }
